fix: avoid wrong route entry

+++ Week2 - integration with MongoDB/src/routes/post.routes.ts
@@ -1,6 +1,7 @@
 import { Resolver } from 'dns';
 import { IncomingMessage, ServerResponse } from 'http';
 import PostController from '../controllers/post.controller';
+import { errorHandle } from '../services';
 
 const postRoute = async (req: IncomingMessage, res: ServerResponse) => {
   const { url, method } = req;
@@ -16,28 +17,18 @@ const postRoute = async (req: IncomingMessage, res: ServerResponse) => {
 
   const id = url?.split('/').pop() ?? '';
 
-  switch (method) {
-    case 'GET':
-      if (url === '/posts') {
-        PostController.getAllPosts(res);
-      } else if (url?.startsWith('/posts/')) {
-        PostController.getPostById(id, res);
-      }
-      break;
-    case 'POST':
-      if (url === '/posts') {
-        PostController.addPost(body, res);
-      }
-    case 'PATCH':
-      if (url?.startsWith('/posts/')) {
-        PostController.updatePost(id, body, res);
-      }
-    case 'DELETE':
-      if (url?.startsWith('/posts/')) {
-        PostController.deletePostById(id, res);
-      }
-    default:
-      break;
+  if (method === 'GET' && url === '/posts') {
+    PostController.getAllPosts(res);
+  } else if (method === 'GET' && url?.startsWith('/posts/')) {
+    PostController.getPostById(id, res);
+  } else if (method === 'POST' && url === '/posts') {
+    PostController.addPost(body, res);
+  } else if (method === 'PATCH' && url?.startsWith('/posts/')) {
+    PostController.updatePost(id, body, res);
+  } else if (method === 'DELETE' && url?.startsWith('/posts/')) {
+    PostController.deletePostById(id, res);
+  } else {
+    errorHandle(res, { data: 'Invalid request' }, 404);
   }
 };
